+++ Changelog.md
@@ -39,6 +39,7 @@
 - Add expressions: spellname[number spellid], componentname[number componentid], componentdata[number componentid], getknownspells[], spelldata[number spellid]
 - Add commands: /ub setmotion, /ub getmotion, /ub clearmotion
 - Add expressions: setmotion[motions, state], getmotion[motion], clearmotion[]
+- Add expression: netclients[string tag?]
 
 ## 0.2.5 (2021-10-06) [UtilityBeltInstaller-0.2.5.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/dd7285b144ef9327bf8041c64841c22f/UtilityBeltInstaller-0.2.5.exe)
 - Add expressions: wobjectgetdoubleprop[], wobjectgetboolprop[], wobjectgetstringprop[], wobjecthasdata[], wobjectrequestdata[]